fix type

+++ src/femr/models/scripts.py
@@ -64,12 +64,12 @@ def train_model() -> None:
     parser.add_argument("--token_dropout", type=float, default=0)
     parser.add_argument("--internal_dropout", type=float, default=0)
     parser.add_argument("--weight_decay", type=float, default=0)
-    parser.add_argument("--max_iter", type=float, default=None)
-    parser.add_argument("--hidden_size", type=float, default=768, help="Transformer hidden size")
-    parser.add_argument("--intermediate_size", type=float, default=3072, help="Transformer intermediate layer size")
-    parser.add_argument("--n_heads", type=float, default=12, help="Transformer # of heads")
-    parser.add_argument("--n_layers", type=float, default=6, help="Transformer # of layers")
-    parser.add_argument("--attention_width", type=float, default=512, help="Transformer attention width.")
+    parser.add_argument("--max_iter", type=int, default=None)
+    parser.add_argument("--hidden_size", type=int, default=768, help="Transformer hidden size")
+    parser.add_argument("--intermediate_size", type=int, default=3072, help="Transformer intermediate layer size")
+    parser.add_argument("--n_heads", type=int, default=12, help="Transformer # of heads")
+    parser.add_argument("--n_layers", type=int, default=6, help="Transformer # of layers")
+    parser.add_argument("--attention_width", type=int, default=512, help="Transformer attention width.")
 
     parser.add_argument(
         "--early_stopping_window_steps",
